Refactor classes and update comments
