+++ src/main/java/seedu/ptman/logic/LogicManager.java
@@ -13,7 +13,7 @@ import seedu.ptman.logic.parser.exceptions.ParseException;
 import seedu.ptman.model.Model;
 import seedu.ptman.model.employee.Employee;
 import seedu.ptman.model.outlet.OutletInformation;
-import seedu.ptman.model.outlet.Shift;
+import seedu.ptman.model.shift.Shift;
 
 /**
  * The main LogicManager of the app.
