+++ src/main/java/seedu/ptman/logic/commands/AddShiftCommand.java
@@ -8,8 +8,8 @@ import static seedu.ptman.logic.parser.CliSyntax.PREFIX_TIME_END;
 import static seedu.ptman.logic.parser.CliSyntax.PREFIX_TIME_START;
 
 import seedu.ptman.logic.commands.exceptions.CommandException;
-import seedu.ptman.model.outlet.Shift;
 import seedu.ptman.model.outlet.exceptions.DuplicateShiftException;
+import seedu.ptman.model.shift.Shift;
 
 /**
  * Adds a shift to PTMan.
