+++ src/main/java/seedu/ptman/logic/commands/ApplyCommand.java
@@ -14,9 +14,9 @@ import seedu.ptman.model.Model;
 import seedu.ptman.model.Password;
 import seedu.ptman.model.employee.Employee;
 import seedu.ptman.model.employee.exceptions.DuplicateEmployeeException;
-import seedu.ptman.model.outlet.Shift;
 import seedu.ptman.model.outlet.exceptions.DuplicateShiftException;
 import seedu.ptman.model.outlet.exceptions.ShiftNotFoundException;
+import seedu.ptman.model.shift.Shift;
 
 /**
  * Registers an employee to a shift identified using their last displayed index from PTMan.
