+++ src/main/java/seedu/ptman/logic/commands/DeleteShiftCommand.java
@@ -9,8 +9,8 @@ import java.util.Objects;
 import seedu.ptman.commons.core.Messages;
 import seedu.ptman.commons.core.index.Index;
 import seedu.ptman.logic.commands.exceptions.CommandException;
-import seedu.ptman.model.outlet.Shift;
 import seedu.ptman.model.outlet.exceptions.ShiftNotFoundException;
+import seedu.ptman.model.shift.Shift;
 
 /**
  * Deletes a shift identified using it's last displayed index in the timetable.
