+++ src/main/java/seedu/ptman/model/ReadOnlyPartTimeManager.java
@@ -3,7 +3,7 @@ package seedu.ptman.model;
 import javafx.collections.ObservableList;
 import seedu.ptman.model.employee.Employee;
 import seedu.ptman.model.outlet.OutletInformation;
-import seedu.ptman.model.outlet.Shift;
+import seedu.ptman.model.shift.Shift;
 import seedu.ptman.model.tag.Tag;
 
 /**
