+++ src/main/java/seedu/ptman/model/shift/UniqueShiftList.java
@@ -1,4 +1,4 @@
-package seedu.ptman.model.outlet;
+package seedu.ptman.model.shift;
 
 import static java.util.Objects.requireNonNull;
 import static seedu.ptman.commons.util.CollectionUtil.requireAllNonNull;
@@ -12,6 +12,7 @@ import seedu.ptman.commons.util.CollectionUtil;
 import seedu.ptman.model.outlet.exceptions.DuplicateShiftException;
 import seedu.ptman.model.outlet.exceptions.ShiftNotFoundException;
 
+//@@author shanwpf
 /**
  * A list of shifts that enforces uniqueness between its elements and does not allow nulls.
  *
