+++ src/test/java/seedu/ptman/logic/commands/AddShiftCommandIntegrationTest.java
@@ -14,9 +14,10 @@ import seedu.ptman.model.ModelManager;
 import seedu.ptman.model.Password;
 import seedu.ptman.model.UserPrefs;
 import seedu.ptman.model.outlet.OutletInformation;
-import seedu.ptman.model.outlet.Shift;
+import seedu.ptman.model.shift.Shift;
 import seedu.ptman.testutil.ShiftBuilder;
 
+//@@author shanwpf
 /**
  * Contains integration tests (interaction with the Model) for {@code AddShiftCommand}.
  */
