+++ src/test/java/seedu/ptman/logic/commands/AddShiftCommandTest.java
@@ -28,13 +28,14 @@ import seedu.ptman.model.employee.Employee;
 import seedu.ptman.model.employee.exceptions.DuplicateEmployeeException;
 import seedu.ptman.model.employee.exceptions.EmployeeNotFoundException;
 import seedu.ptman.model.outlet.OutletInformation;
-import seedu.ptman.model.outlet.Shift;
 import seedu.ptman.model.outlet.exceptions.DuplicateShiftException;
 import seedu.ptman.model.outlet.exceptions.NoOutletInformationFieldChangeException;
 import seedu.ptman.model.outlet.exceptions.ShiftNotFoundException;
+import seedu.ptman.model.shift.Shift;
 import seedu.ptman.model.tag.Tag;
 import seedu.ptman.testutil.ShiftBuilder;
 
+//@@author shanwpf
 public class AddShiftCommandTest {
 
     @Rule
