+++ src/test/java/seedu/ptman/logic/parser/AddShiftCommandParserTest.java
@@ -3,8 +3,8 @@ package seedu.ptman.logic.parser;
 import static seedu.ptman.commons.core.Messages.MESSAGE_INVALID_COMMAND_FORMAT;
 import static seedu.ptman.logic.commands.CommandTestUtil.CAPACITY_DESC_1;
 import static seedu.ptman.logic.commands.CommandTestUtil.CAPACITY_DESC_2;
-import static seedu.ptman.logic.commands.CommandTestUtil.DATE_DESC_MONDAY;
-import static seedu.ptman.logic.commands.CommandTestUtil.DATE_DESC_TUESDAY;
+import static seedu.ptman.logic.commands.CommandTestUtil.DATE_DESC_12MAR;
+import static seedu.ptman.logic.commands.CommandTestUtil.DATE_DESC_13MAR;
 import static seedu.ptman.logic.commands.CommandTestUtil.INVALID_CAPACITY_DESC;
 import static seedu.ptman.logic.commands.CommandTestUtil.INVALID_DATE_DESC;
 import static seedu.ptman.logic.commands.CommandTestUtil.INVALID_TIME_END_DESC;
@@ -25,12 +25,13 @@ import static seedu.ptman.logic.parser.CommandParserTestUtil.assertParseSuccess;
 import org.junit.Test;
 
 import seedu.ptman.logic.commands.AddShiftCommand;
-import seedu.ptman.model.outlet.Capacity;
-import seedu.ptman.model.outlet.Date;
-import seedu.ptman.model.outlet.Shift;
-import seedu.ptman.model.outlet.Time;
+import seedu.ptman.model.shift.Capacity;
+import seedu.ptman.model.shift.Date;
+import seedu.ptman.model.shift.Shift;
+import seedu.ptman.model.shift.Time;
 import seedu.ptman.testutil.ShiftBuilder;
 
+//@@author shanwpf
 public class AddShiftCommandParserTest {
     private AddShiftCommandParser parser = new AddShiftCommandParser();
 
@@ -40,23 +41,23 @@ public class AddShiftCommandParserTest {
                 .withEndTime(VALID_TIME_END_8PM).withCapacity(VALID_CAPACITY_1).build();
 
         // whitespace only preamble
-        assertParseSuccess(parser, PREAMBLE_WHITESPACE + DATE_DESC_MONDAY + TIME_START_DESC_10AM + TIME_END_DESC_8PM
+        assertParseSuccess(parser, PREAMBLE_WHITESPACE + DATE_DESC_12MAR + TIME_START_DESC_10AM + TIME_END_DESC_8PM
                 + CAPACITY_DESC_1, new AddShiftCommand(expectedShift));
 
-        // multiple days - last day accepted
-        assertParseSuccess(parser, DATE_DESC_TUESDAY + DATE_DESC_MONDAY + TIME_START_DESC_10AM + TIME_END_DESC_8PM
+        // multiple dates - last day accepted
+        assertParseSuccess(parser, DATE_DESC_13MAR + DATE_DESC_12MAR + TIME_START_DESC_10AM + TIME_END_DESC_8PM
                 + CAPACITY_DESC_1, new AddShiftCommand(expectedShift));
 
         // multiple start times - last start time accepted
-        assertParseSuccess(parser, DATE_DESC_MONDAY + TIME_START_DESC_12PM + TIME_START_DESC_10AM + TIME_END_DESC_8PM
+        assertParseSuccess(parser, DATE_DESC_12MAR + TIME_START_DESC_12PM + TIME_START_DESC_10AM + TIME_END_DESC_8PM
                 + CAPACITY_DESC_1, new AddShiftCommand(expectedShift));
 
         // multiple end times - last end time accepted
-        assertParseSuccess(parser, DATE_DESC_MONDAY + TIME_START_DESC_10AM +  TIME_END_DESC_10PM + TIME_END_DESC_8PM
+        assertParseSuccess(parser, DATE_DESC_12MAR + TIME_START_DESC_10AM +  TIME_END_DESC_10PM + TIME_END_DESC_8PM
                 + CAPACITY_DESC_1, new AddShiftCommand(expectedShift));
 
         // multiple capacities - last capacity accepted
-        assertParseSuccess(parser, DATE_DESC_MONDAY + TIME_START_DESC_10AM +  TIME_END_DESC_8PM
+        assertParseSuccess(parser, DATE_DESC_12MAR + TIME_START_DESC_10AM +  TIME_END_DESC_8PM
                 + CAPACITY_DESC_2 + CAPACITY_DESC_1, new AddShiftCommand(expectedShift));
 
     }
@@ -65,18 +66,18 @@ public class AddShiftCommandParserTest {
     public void parse_compulsoryFieldMissing_failure() {
         String expectedMessage = String.format(MESSAGE_INVALID_COMMAND_FORMAT, AddShiftCommand.MESSAGE_USAGE);
 
-        // missing day prefix
+        // missing date prefix
         assertParseFailure(parser,  VALID_DATE_12MAR + TIME_START_DESC_10AM + TIME_END_DESC_8PM
                 + CAPACITY_DESC_1, expectedMessage);
 
         // missing start time prefix
-        assertParseFailure(parser,  DATE_DESC_MONDAY + VALID_TIME_START_10AM + TIME_END_DESC_8PM
+        assertParseFailure(parser,  DATE_DESC_12MAR + VALID_TIME_START_10AM + TIME_END_DESC_8PM
                 + CAPACITY_DESC_1, expectedMessage);
         // missing end time prefix
-        assertParseFailure(parser,  DATE_DESC_MONDAY + TIME_START_DESC_10AM + VALID_TIME_END_8PM
+        assertParseFailure(parser,  DATE_DESC_12MAR + TIME_START_DESC_10AM + VALID_TIME_END_8PM
                 + CAPACITY_DESC_1, expectedMessage);
         // missing capacity prefix
-        assertParseFailure(parser,  DATE_DESC_MONDAY + TIME_START_DESC_10AM + TIME_END_DESC_8PM
+        assertParseFailure(parser,  DATE_DESC_12MAR + TIME_START_DESC_10AM + TIME_END_DESC_8PM
                 + VALID_CAPACITY_1, expectedMessage);
 
         // all prefixes missing
@@ -86,28 +87,28 @@ public class AddShiftCommandParserTest {
 
     @Test
     public void parse_invalidValue_failure() {
-        // invalid day
+        // invalid date
         assertParseFailure(parser, INVALID_DATE_DESC + TIME_START_DESC_10AM + TIME_END_DESC_8PM
                 + CAPACITY_DESC_1, Date.MESSAGE_DATE_CONSTRAINTS);
 
         // invalid start time
-        assertParseFailure(parser, DATE_DESC_MONDAY + INVALID_TIME_START_DESC + TIME_END_DESC_8PM
+        assertParseFailure(parser, DATE_DESC_12MAR + INVALID_TIME_START_DESC + TIME_END_DESC_8PM
                         + CAPACITY_DESC_1, Time.MESSAGE_TIME_CONSTRAINTS);
 
         // invalid end time
-        assertParseFailure(parser, DATE_DESC_MONDAY + TIME_START_DESC_10AM + INVALID_TIME_END_DESC
+        assertParseFailure(parser, DATE_DESC_12MAR + TIME_START_DESC_10AM + INVALID_TIME_END_DESC
                         + CAPACITY_DESC_1, Time.MESSAGE_TIME_CONSTRAINTS);
 
         // invalid capacity
-        assertParseFailure(parser, DATE_DESC_MONDAY + TIME_START_DESC_10AM + TIME_END_DESC_8PM
+        assertParseFailure(parser, DATE_DESC_12MAR + TIME_START_DESC_10AM + TIME_END_DESC_8PM
                         + INVALID_CAPACITY_DESC, Capacity.MESSAGE_CAPACITY_CONSTRAINTS);
 
         // two invalid values, only first invalid value reported
-        assertParseFailure(parser, DATE_DESC_MONDAY + TIME_START_DESC_10AM + INVALID_TIME_END_DESC
+        assertParseFailure(parser, DATE_DESC_12MAR + TIME_START_DESC_10AM + INVALID_TIME_END_DESC
                         + INVALID_CAPACITY_DESC, Time.MESSAGE_TIME_CONSTRAINTS);
 
         // non-empty preamble
-        assertParseFailure(parser, PREAMBLE_NON_EMPTY + DATE_DESC_MONDAY + TIME_START_DESC_10AM + TIME_END_DESC_8PM
+        assertParseFailure(parser, PREAMBLE_NON_EMPTY + DATE_DESC_12MAR + TIME_START_DESC_10AM + TIME_END_DESC_8PM
                         + INVALID_CAPACITY_DESC,
                 String.format(MESSAGE_INVALID_COMMAND_FORMAT, AddShiftCommand.MESSAGE_USAGE));
     }
