+++ src/test/java/seedu/ptman/logic/parser/DeleteShiftCommandParserTest.java
@@ -9,13 +9,7 @@ import org.junit.Test;
 
 import seedu.ptman.logic.commands.DeleteShiftCommand;
 
-/**
- * As we are only doing white-box testing, our test cases do not cover path variations
- * outside of the DeleteShiftCommand code. For example, inputs "1" and "1 abc" take the
- * same path through the DeleteShiftCommand, and therefore we test only one of them.
- * The path variation for those two cases occur inside the ParserUtil, and
- * therefore should be covered by the ParserUtilTest.
- */
+//@@author shanwpf
 public class DeleteShiftCommandParserTest {
 
     private DeleteShiftCommandParser parser = new DeleteShiftCommandParser();
