+++ src/test/java/seedu/ptman/logic/parser/UnapplyCommandParserTest.java
@@ -13,13 +13,7 @@ import org.junit.Test;
 import seedu.ptman.logic.commands.UnapplyCommand;
 import seedu.ptman.model.Password;
 
-/**
- * As we are only doing white-box testing, our test cases do not cover path variations
- * outside of the UnapplyCommand code. For example, inputs "1" and "1 abc" take the
- * same path through the UnapplyCommand, and therefore we test only one of them.
- * The path variation for those two cases occur inside the ParserUtil, and
- * therefore should be covered by the ParserUtilTest.
- */
+//@@author shanwpf
 public class UnapplyCommandParserTest {
 
     private UnapplyCommandParser parser = new UnapplyCommandParser();
