+++ src/test/java/seedu/ptman/testutil/ShiftUtil.java
@@ -6,29 +6,29 @@ import static seedu.ptman.logic.parser.CliSyntax.PREFIX_TIME_END;
 import static seedu.ptman.logic.parser.CliSyntax.PREFIX_TIME_START;
 
 import seedu.ptman.logic.commands.AddShiftCommand;
-import seedu.ptman.model.outlet.Shift;
+import seedu.ptman.model.shift.Shift;
 
+//@@author shanwpf
 /**
  * A utility class for Shift.
  */
 public class ShiftUtil {
-    public static final String DEFAULT_PASSWORD = "DEFAULT1";
     /**
-     * Returns an add command string for adding the {@code employee}.
+     * Returns an addshift command string for adding the {@code shift}.
      */
     public static String getAddShiftCommand(Shift shift) {
         return AddShiftCommand.COMMAND_WORD + " " + getShiftDetails(shift);
     }
 
     /**
-     * Returns an aliased add command string for adding the {@code employee}.
+     * Returns an aliased addshift command string for adding the {@code shift}.
      */
     public static String getAliasedAddShiftCommand(Shift shift) {
         return AddShiftCommand.COMMAND_ALIAS + " " + getShiftDetails(shift);
     }
 
     /**
-     * Returns the part of command string for the given {@code employee}'s details.
+     * Returns the part of command string for the given {@code shift}'s details.
      */
     public static String getShiftDetails(Shift shift) {
         StringBuilder sb = new StringBuilder();
